requirements.txt
absl-py==2.3.0
adal==1.2.7
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
applicationinsights==0.11.10
arch==5.6.0
argcomplete==3.6.0
asgiref==3.9.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
azure-ai-ml==1.28.1
azure-common==1.1.28
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-graphrbac==0.61.2
azure-identity==1.23.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.5.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==28.1.0
azure-mgmt-resource==23.3.0
azure-mgmt-storage==22.0.0
azureml-automl-core==1.59.0
azureml-automl-runtime==1.59.0
azureml-core==1.59.0.post1
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.5
azureml-dataset-runtime==1.59.0
azureml-defaults==1.59.0
azureml-inference-server-http==1.4.0
azureml-interpret==1.59.0
azureml-mlflow==1.59.0.post1
azureml-pipeline-core==1.59.0
azureml-responsibleai==1.59.0
azureml-telemetry==1.59.0
azureml-train-automl==1.59.0
azureml-train-automl-client==1.59.0
azureml-train-automl-runtime==1.59.0
azureml-train-core==1.59.0
azureml-training-tabular==1.59.0
azureml-train-restclients-hyperdrive==1.59.0
azure-monitor-opentelemetry==1.6.10
azure-monitor-opentelemetry-exporter==1.0.0b39
azureopenai==0.0.1
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
azure-storage-file-share==12.21.0
azure-storage-queue==12.12.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.3.0
blinker==1.9.0
bokeh==2.4.3
boto3==1.20.19
botocore==1.23.19
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
cmdstanpy==1.2.5
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.0
convertdate==2.4.0
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
dask==2023.2.0
databricks-sdk==0.45.0
dataclasses==0.6
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dice-ml==0.11
dill==0.3.9
distributed==2023.2.0
distro==1.9.0
docker==7.1.0
dotnetcore2==3.1.23
econml==0.15.1
entrypoints==0.4
ephem==4.2
erroranalysis==0.5.5
exceptiongroup==1.2.2
executing==2.2.0
fairlearn==0.7.0
fastapi==0.116.0
fire==0.7.0
fixedint==0.1.6
Flask==3.0.3
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.7.0
fsspec==2025.3.0
fusepy==3.0.1
gast==0.6.0
gensim==4.3.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
googleapis-common-protos==1.69.1
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
grpcio==1.72.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
holidays==0.76
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==7.2.1
importlib_resources==6.4.0
inference-schema==1.8
interpret_community==0.31.0
interpret-core==0.5.0
ipykernel==6.29.5
ipython==8.18.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
jsonpickle==4.0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kapre==0.3.7
keras==3.10.0
keras2onnx==1.6.0
kiwisolver==1.4.7
knack==0.12.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
lightgbm==3.2.1
llvmlite==0.43.0
locket==1.0.0
LunarCalendar==0.0.9
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
ml_wrappers==0.5.6
mlflow==3.1.1
mlflow-skinny==3.1.1
mltable==1.6.1
moviepy==2.2.1
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msgpack==1.1.0
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.4.4
namex==0.1.0
narwhals==1.46.0
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
networkx==3.2.1
noisereduce==3.0.3
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
onnx==1.17.0
onnxconverter-common==1.13.0
onnxmltools==1.11.2
onnxruntime==1.17.3
openai==1.91.0
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
openl3==0.4.2
opentelemetry-api==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-dbapi==0.51b0
opentelemetry-instrumentation-django==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-flask==0.51b0
opentelemetry-instrumentation-psycopg2==0.51b0
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-instrumentation-urllib==0.51b0
opentelemetry-instrumentation-urllib3==0.51b0
opentelemetry-instrumentation-wsgi==0.51b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==1.5.3
paramiko==3.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
pip==25.1.1
pkginfo==1.12.1.2
platformdirs==4.3.6
plotly==6.2.0
pmdarima==1.8.5
pooch==1.8.2
portalocker==2.10.1
proglog==0.1.12
prompt_toolkit==3.0.50
propcache==0.3.2
property-cached==1.6.4
prophet==1.1.4
protobuf==5.28.3
proto-plus==1.26.0
psutil==5.9.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==14.0.2
pyasn1_modules==0.4.1
pyasn1==0.6.1
pycparser==2.22
pydantic_core==2.23.4
pydantic==2.9.2
pydantic-settings==2.8.1
pydash==8.0.5
Pygments==2.19.1
PyJWT==2.10.1
PyMeeus==0.5.12
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.2.1
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.2.1
raiutils==0.4.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.2.2
responsibleai==0.36.0
rich==14.0.0
rpds-py==0.23.1
rsa==4.9
s3transfer==0.5.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
SecretStorage==3.3.3
semver==2.13.0
setuptools==58.1.0
shap==0.43.0
six==1.17.0
skl2onnx==1.15.0
sklearn-pandas==1.7.0
slicer==0.0.7
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
sparse==0.15.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
stanio==0.5.1
starlette==0.46.2
statsmodels==0.13.5
strictyaml==1.7.3
sympy==1.14.0
tabulate==0.9.0
tblib==3.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.35.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.16.0
xgboost==1.5.2
yarl==1.20.1
zict==3.0.0
zipp==3.21.0
